# Use f-strings instead of format