--- a/cclib/bridge/cclib2openbabel.py
+++ b/cclib/bridge/cclib2openbabel.py
@@ -77,7 +77,7 @@ def readfile(fname, format):
         obc.ReadFile(mol, fname)
         return makecclib(mol)
     else:
-        print("Unable to load the %s reader from OpenBabel." % format)
+        print(f"Unable to load the {format} reader from OpenBabel.")
         return {}