--- a/cclib/bridge/cclib2pyquante.py
+++ b/cclib/bridge/cclib2pyquante.py
@@ -36,7 +36,7 @@ def makepyquante(data):
     if missing:
         missing = " ".join(missing)
         raise MissingAttributeError(
-            "Could not create pyquante molecule due to missing attribute: {}".format(missing)
+            f"Could not create pyquante molecule due to missing attribute: {missing}"
         )
 
     # In pyquante2, molecular geometry is specified in a format of: