--- a/cclib/io/xyzwriter.py
+++ b/cclib/io/xyzwriter.py
@@ -88,15 +88,15 @@ class XYZ(filewriter.Writer):
         else:
             geometry_num = index + 1
         if self.jobfilename is not None:
-            comment = "{}: Geometry {}".format(self.jobfilename, geometry_num)
+            comment = f"{self.jobfilename}: Geometry {geometry_num}"
         else:
-            comment = "Geometry {}".format(geometry_num)
+            comment = f"Geometry {geometry_num}"
         # Wrap the geometry number part of the comment in square brackets,
         # prefixing it with one previously parsed if it existed.
         if existing_comment:
-            comment = "{} [{}]".format(existing_comment, comment)
+            comment = f"{existing_comment} [{comment}]"
         else:
-            comment = "[{}]".format(comment)
+            comment = f"[{comment}]"
 
         atom_template = '{:3s} {:15.10f} {:15.10f} {:15.10f}'
         block = []