--- a/cclib/method/calculationmethod.py
+++ b/cclib/method/calculationmethod.py
@@ -44,7 +44,7 @@ class Method:
         self.loglevel = loglevel
         self.logname = logname
         self._check_required_attributes()
-        self.logger = logging.getLogger('%s %s' % (self.logname, self.data))
+        self.logger = logging.getLogger(f"{self.logname} {self.data}")
         self.logger.setLevel(self.loglevel)
         self.logformat = "[%(name)s %(levelname)s] %(message)s"
         handler = logging.StreamHandler(sys.stdout)
@@ -58,4 +58,5 @@ class Method:
         if missing:
             missing = ' '.join(missing)
             raise MissingAttributeError(
-                'Could not parse required attributes to use method: ' + missing)
+                f"Could not parse required attributes to use method: {missing}"
+            )