--- a/cclib/method/cspa.py
+++ b/cclib/method/cspa.py
@@ -25,11 +25,11 @@ class CSPA(Population):
 
     def __str__(self):
         """Return a string representation of the object."""
-        return "CSPA of %s" % (self.data)
+        return f"CSPA of {self.data}"
 
     def __repr__(self):
         """Return a representation of the object."""
-        return 'CSPA("%s")' % (self.data)
+        return f'CSPA("{self.data}")'
 
     def calculate(self, indices=None, fupdate=0.05):
         """Perform the C squared population analysis.