--- a/cclib/method/fragments.py
+++ b/cclib/method/fragments.py
@@ -25,11 +25,11 @@ class FragmentAnalysis(Method):
 
     def __str__(self):
         """Return a string representation of the object."""
-        return "Fragment molecule basis of %s" % (self.data)
+        return f"Fragment molecule basis of {self.data}"
 
     def __repr__(self):
         """Return a representation of the object."""
-        return 'Fragment molecular basis("%s")' % (self.data)
+        return f'Fragment molecular basis("{self.data}")'
 
     def calculate(self, fragments, cupdate=0.05):
 
@@ -54,9 +54,9 @@ class FragmentAnalysis(Method):
             #assign fonames based on fragment name and MO number
             for i in range(fragments[j].nbasis):
                 if hasattr(fragments[j],"name"):
-                    self.fonames.append("%s_%i"%(fragments[j].name,i+1))
+                    self.fonames.append(f"{fragments[j].name}_{int(i + 1)}")
                 else:
-                    self.fonames.append("noname%i_%i"%(j,i+1))
+                    self.fonames.append(f"noname{int(j)}_{int(i + 1)}")
 
         nBasis = self.data.nbasis
         nAlpha = self.data.homos[0] + 1