--- a/cclib/method/hirshfeld.py
+++ b/cclib/method/hirshfeld.py
@@ -56,11 +56,11 @@ class Hirshfeld(Stockholder):
 
     def __str__(self):
         """Return a string representation of the object."""
-        return "Hirshfeld charges of {}".format(self.data)
+        return f"Hirshfeld charges of {self.data}"
 
     def __repr__(self):
         """Return a representation of the object."""
-        return "Hirshfeld({})".format(self.data)
+        return f"Hirshfeld({self.data})"
 
     def _check_required_attributes(self):
         super()._check_required_attributes()