--- a/manifest.py
+++ b/manifest.py
@@ -23,6 +23,6 @@ for folder in folders:
 
 for f in files:
     if not os.path.isfile(f):
-        print("%s does not exist" % f)
+        print(f"{f} does not exist")
 
 print("\n".join(files), file=open("MANIFEST", "w"))