I did some profiling on the regression test set and realized `.format()` calls incur a considerable performance penalty (~10% of the whole runtime for the regression test set!). So I changed everything to f-strings (only Python 3.6+ though).

## Profiling

```
$ git checkout upstream/master
$ python3 -m cProfile -o upstream.pstats -m test.regression
...
$ python3 -m pstats upstream.pstats
Welcome to the profile statistics browser.
upstream.pstats% strip
upstream.pstats% sort time
upstream.pstats% stats 10
Fri Mar 25 21:47:13 2022    upstream.pstats

         332588011 function calls (327605908 primitive calls) in 135.552 seconds

   Ordered by: internal time
   List reduced from 4242 to 10 due to restriction <10>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
  1562678   36.925    0.000   63.220    0.000 gaussianparser.py:170(extract)
  1783834   12.959    0.000   25.040    0.000 orcaparser.py:79(extract)
 43724395    8.860    0.000    8.860    0.000 {method 'format' of 'str' objects}
 91811535    8.855    0.000    8.855    0.000 {method 'lower' of 'str' objects}
   607366    4.901    0.000    7.767    0.000 adfparser.py:99(extract)
      531    4.820    0.009  132.533    0.250 logfileparser.py:267(parse)
  9789736    4.400    0.000    7.429    0.000 logfileparser.py:90(next)
   404827    3.967    0.000    6.780    0.000 gamessparser.py:80(extract)
 45795664    3.900    0.000    3.900    0.000 {method 'strip' of 'str' objects}
14473526/9965014    3.130    0.000    6.701    0.000 {built-in method builtins.next}

```

[The profile output.](https://github.com/cclib/cclib/files/8354420/upstream.zip)

## Actual benchmarks

### Baseline

upstream `master` branch (f4f95c800cd5681adb2ba693394aa571372544b6):

```
$ git checkout upstream/master
$ hyperfine 'python3 -m test.regression' --warmup 2 --min-runs 7
Benchmark 1: python3 -m test.regression
  Time (mean ± σ):     83.120 s ±  0.575 s    [User: 83.487 s, System: 2.399 s]
  Range (min … max):   82.359 s … 83.677 s    7 runs

```

### This pull request

`use-f-str`:

```
$ git checkout use-f-str
$ hyperfine 'python3 -m test.regression' --warmup 2 --min-runs 7
Benchmark 1: python3 -m test.regression
  Time (mean ± σ):     77.278 s ±  0.528 s    [User: 77.663 s, System: 2.376 s]
  Range (min … max):   76.615 s … 78.074 s    7 runs

```

Which is an improvement of 7.6% (±1.0%) overall.

## Reproducing

```console
$ cargo install hyperfine  # benchmarking tool, requires Rust
$ hyperfine -V
hyperfine 1.13.0
$ git clone git@github.com:cclib/cclib.git
$ cd cclib/data
$ source regression_download.sh
$ cd ..
$ python3 -m test.regression
...
Total: 531   Failed: 0  Errors: 0

WARNING: There are 1 orphaned regression test functions.
Please make sure these function names correspond to regression files:
testQChem_QChem5_1_old_final_print_1_out

```