@berquist 2022-03-27
https://github.com/cclib/cclib/runs/5711973999?check_suite_focus=true#step:7:632 really?