@schneiderfelipe 2022-03-28
By the way, I use [flynt](https://github.com/ikamensh/flynt) for most of this 😆